move screen file
